refactor(ui): FlexStack - convert to typescript
